Data is not handled by DataSet, taken off emotion.
Wrote the code which maps the various emotions.txt files into something
more usable.  DataSet now takes one file for each emotion and combines
them to create three sets of data: the training set, the development set
and the test set.  The proportion of sentences per set is 8:1:1
respectively

The training set is an array where each element is of the form
[sentence, classifiction]

The dev and test sets are flat arrays, with each element a different
sentence.

The sentences are combined and shuffled, before being split into their
correct proportions.

Sentences in each set appear only in that set.

In the event that a whole number of sentences cannot be found, the
number is rounded up.

--

Splitting the data into the correct proportions is currently tested by a
single unit test.  The other behaviour described above is currently
not explicitly tested.