Word count and probabilities moved to Classifier, refactored DataSet.
DataSet now has a simpler api, always returns a set with #set.

You can choose which set is returned (i.e. training, dev or test) by
calling use_set and the name of the set.

Probabilities are calculated on Classifier, and the data structures for
holding the wordcounts have been changed slightly.  There are now two
hashes: one to store the counts for each word in the dataset, and
one to store the cumulative counts for each sentiment.

I *think* this completes the 'training' portion of the Bayes classifier.
What remains is the code to run the simple classifier against the
devset to see whether it produces results. This is what was done for NLP
practical 1.  Perhaps unit tests for this can be 'it must classify with
70 accuracy or greater' or something.

Next up:

* Expand the classifier to go beyond unigrams and use ngrams.
* Smoothing + backoff, i.e. fallback to lower-order ngrams. This
  requires implementing smoothing e.g. as described in Katz backoff.

Test quality has gone downhill, as has overall code quality as use of TDD is
becoming less disciplined.